# packages/api-contracts/openapi.yaml
...
openapi: 3.0.0
paths:
...
  /api/v1/wishlist:
    get:
      description: Requires authentication. Each item carries the price frozen at save time; currentPrice is the live cached price, or null when not currently cached.
      operationId: WishlistController_list_v1
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WishlistItemResponseDto'
                  meta:
                    $ref: '#/components/schemas/WishlistMetaDto'
                required:
                  - success
                  - data
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      security:
        - bearer: []
      summary: List the saved products of the signed-in user
      tags:
        - wishlist
    post:
      description: Requires authentication. Rejected with 422 once the per-user cap (meta.max on GET) is reached.
      operationId: WishlistController_add_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  minLength: 1
                token:
                  type: string
                  default: ''
                title:
                  type: string
                  minLength: 1
                imageUrl:
                  type: string
                  default: ''
                price:
                  type: number
                  minimum: 0
                platform:
                  type: string
                  default: ''
                productUrl:
                  type: string
                  default: ''
              required:
                - productId
                - title
                - price
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/WishlistItemResponseDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '422':
          description: BUSINESS_RULE_ERROR — the request was well-formed but violates a rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      security:
        - bearer: []
      summary: Save a product to the wishlist
      tags:
        - wishlist
    delete:
      description: Requires authentication.
      operationId: WishlistController_clearAll_v1
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/OkResultDto'
                required:
                  - success
                  - data
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      security:
        - bearer: []
      summary: Clear the whole wishlist
      tags:
        - wishlist
  /api/v1/wishlist/{productId}:
    delete:
      description: Requires authentication. Idempotent — removing something absent still reports ok.
      operationId: WishlistController_remove_v1
      parameters:
        - name: productId
          required: true
          in: path
          description: Product id as stored on the wishlist item.
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/OkResultDto'
                required:
                  - success
                  - data
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      security:
        - bearer: []
      summary: Remove one saved product
      tags:
        - wishlist
...
info:
  title: Omnisift API
  description: Product Intelligence Platform — REST API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - AUTHENTICATION_ERROR
        - AUTHORIZATION_ERROR
        - NOT_FOUND
        - PROVIDER_ERROR
        - DATABASE_ERROR
        - BUSINESS_RULE_ERROR
        - INTERNAL_ERROR
      description: Stable machine-readable code. Safe to branch on.
    ApiErrorDetailDto:
      type: object
      properties:
        code:
          description: Stable machine-readable code. Safe to branch on.
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Safe, user-facing message. Never contains a stack trace or internal detail.
      required:
        - code
        - message
    ApiErrorResponseDto:
      type: object
      properties:
        success:
          type: boolean
          example: false
          description: Always false on an error response.
        error:
          $ref: '#/components/schemas/ApiErrorDetailDto'
        requestId:
          type: string
          nullable: true
          description: Correlation id echoed from the x-request-id request header, or null.
      required:
        - success
        - error
        - requestId
...
    OkResultDto:
      type: object
      properties:
        ok:
          type: boolean
          example: true
      required:
        - ok
...
    WishlistItemResponseDto:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        token:
          type: string
        title:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        platform:
          type: string
        productUrl:
          type: string
        currentPrice:
          type: number
          nullable: true
      required:
        - id
        - productId
        - token
        - title
        - imageUrl
        - price
        - platform
        - productUrl
        - currentPrice
    WishlistMetaDto:
      type: object
      properties:
        count:
          type: number
        max:
          type: number
          description: Maximum items a user may save, from the app_config.wishlist_max_items runtime dial.
      required:
        - count
        - max
...